Move unit test for overlap algo to new class UnitTestRuleOverlapAlgo
